+++ src/main/java/de/adorsys/keycloak/config/service/state/StateService.java
@@ -65,7 +65,7 @@ public class StateService {
     }
 
     public void doImport(RealmImport realmImport) {
-        if (!importConfigProperties.isState()) {
+        if (!importConfigProperties.getRemoteState().isEnabled()) {
             return;
         }